These aren't part of the XDG standard
Although they look part of it, they're not, and actually use a different
mechanism (see user-dirs.dirs) which requires its own
shell-like-but-not-quite parsing as they're not environment variables.

So for example the current tests will assume /home/earnest/Desktop is
correct for XDG_DESKTOP_HOME when it's actually defined to be

    XDG_DESKTOP_DIR="$HOME/"

in my XDG_CONFIG_HOME/user-dirs.dirs file.  (The quotes and trailing '/'
is actually required.)